requirements.txt
pillow
requests
pytz
Babel
python-dateutil
aiohttp
click
cryptocompare
pathlib
